stm32/hrtim: move Instance and Prescaler to module root
